Compute latency

+++ media_player.py
@@ -403,7 +403,7 @@ class MediaPlayer():
                     print(f'Second drift = {drift}. Synching...')
                     self.network_latency = abs(drift)
                 elif self.sync_count > 5: # latency is a moving average
-                    self.network_latency = drift * 0.1 + self.network_latency * 0.9
+                    self.network_latency = int(drift * 0.1 + self.network_latency * 0.9)
                     print(f'Latency = {self.network_latency}')
                 self.sync_count += 1
